+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All Notable changes to `League\Uri\Interfaces` will be documented in this file
 
 - `QueryString::compose` and `QueryString::composeFromValue` a userland improved `http_build_query` method.
 - `QueryExtractMode` and `QueryComposeMode` Enum to improve query parsing and building.
+- `StringCoercionMode` to normalize type coercion
 - Added support for `BackedEnum`
 - `FragmentDirective::toFragmentValue`
 - `DomainHostInteraface::isSubdomainOf`
@@ -21,6 +22,7 @@ All Notable changes to `League\Uri\Interfaces` will be documented in this file
 
 - Fix issue [#177](https://github.com/thephpleague/uri-src/issues/177) improved userInfo encoding.
 - Fix `UriString::resolve` method.
+- Fix Type coercion when using the QueryString
 
 ### Deprecated